Add files via upload
Next, click on the "Windows Logs" folder in the far-left column. Then, select "Application" to view the application logs, where the logs for this lab are located.
Scroll through the logs to find the five logs that you need to fix for this lab. They should have a time that's close to when you started the lab, and the "Source" field should be "Qwiklab". To view details of a log entry, click on it in Event Viewer, and a message window will open at the bottom of the Event Viewer window.