Search and delete files
This log is warning you about a large file that's taking up disk space, but it doesn't specify the file name. On Windows, you can find large files using the File Explorer. Open it and navigate to your C:\ directory, then click the search bar at the top right.
click the "Size" dropdown and choose "Gigantic" to start a search for files over 128 MB.

Now that you know which file is causing the low-memory error, you can delete it by right-clicking the file in the search results, and selecting "delete". You'll need to provide admin privileges to delete it. When this popup appears, click "continue"

Files over a certain size don't fit in the Windows "Recycle Bin", and can only be permanently deleted. After clicking "Continue", you'll see this message, indicating that the file is too large to recycle. Press "Yes" to delete permanently.

The file will delete, fixing the log issue! Note that the log will still remain in Event Viewer, because logs aren't deleted when the issues that caused them are removed